--- app.js
@@ -1,3 +1,0 @@
-const a = 1;
-const obj = {}
-console.log(obj['ADMIN'] = 'glist')
